prototype/server/requirements.txt
Flask==1.1.2
python-dotenv==0.15.0
PyYAML==5.3.1
requests==2.24.0